Create JavaScriptDDOS
This code attempts to send a very large number of HTTP requests to the specified server (serverIp) by creating and loading image objects in a loop. It does not take any action upon successful requests but logs an error message to the console if any of the requests fail to load. The purpose of such code could be to stress-test or flood a server with requests, although it's important to note that sending a very large number of requests to a server can be considered a form of network abuse and should only be done responsibly and with proper authorization.